requirements.txt
python-dotenv==1.0.1
requests==2.32.3
djangorestframework==3.15.2
django==4.2.17
flake8==7.1.1
flake8-isort==6.1.1
flake8-quotes==3.4.0
drf-spectacular
gunicorn==20.1.0
celery==5.4.0
django-celery-beat==2.7.0
django-redis==5.4.0
psycopg2==2.9.9
pytest-django==4.5.2
pytest==7.1.3
django-cors-headers